Catch problematic tasks that fail to launch.
Wrapped the Popen call for tasks in a try/except block so that tasks
that cannot even be launched will be noted.

If the exception is raised and it contains an error number, we report
that as the return code, otherwise "200", which is not currently used
as an errno. We also convert the str representation of the exception
into an "uncaught" stderr message. The PID in this case is set to -1.

Added an example that prompted this change: a task with ~320K bytes.