Add dynamic execution contexts
Extensive reworking to allow execution contexts to join and leave a disBatch run.

dbUtil.sh is a template shell script to add execution contexts to the disBatch run. Each disBatch run uses this to generate a script customized to the run. One can, for example, simply submit this script via sbatch to add nodes to the run. Or invoke with '-s localhost:12' to add 12 cores from the current machine.

With the argument '--mon' it starts up a simple ASCII control monitor and control UI.

dbMon.py implements this UI.