Correct GPU gres issue with SLURM.
In order for the engine to see all of the allocated gres resources, it
appears that srun has to request a full set of tasks.

To avoid starting more than one engine, all but one task has to exit,
so srun starts a bash command that exits for all but SLURM_LOCALID==0.

Not entirely sure how the underlying plumbing works (e.g., cgroup's
role), so not entirely sure that this works either.