Fixed issue with resume runs.
The TaskCmd field of a TaskReport needs to be a repr().

parseStatusFiles assumes this is the case when parsing a line from a
status file that is being read in order to resume a run.

Going the other way, OutData and ErrData are repr()s and need to be
parsed with literal_eval() when a status file is read in.

TaskReport code has been changed to use mnemonic field names.

JSON encoding is now specified when adding a json dump of a TaskReport to KVS.